(tests): Implement code refactor
- add travis.yml file
- install react test renderer
- link codeclimate
- add test in homeview
- update README file with badge links

[Finishes #177907762]
