+++ .travis.yml
@@ -1,0 +1,17 @@
+language: node_js
+node_js:
+  - "stable"
+cache:
+  directories:
+    - node_modules
+before_install:
+   - npm i -g npm@version-7.11.1
+install:
+  - npm
+cache:
+  npm: true
+  directories:
+    - node_modules
+script:
+  - npm test --coverage
+  - npm run build
