+++ client/src/components/Home/HomeView/__tests__/HomeView.spec.js
@@ -1,0 +1,11 @@
+import React from "react";
+import renderer from "react-test-renderer";
+import {HomeView} from "../HomeView";
+
+describe("AdminDisplay", () => {
+  it("snapshot renders correctly", () => {
+    const component = renderer.create(<HomeView />);
+    const tree = component.toJSON();
+    expect(tree).toMatchSnapshot();
+  });
+});
